FileSplitter.txt
As website developers we often write code to build a new page, add a new feature, or to simply maintain some existing libraries. Then there are the times when we need to solve one or more problems. This program was written to solve a major problem.

I needed a way to track changes made to any number of fields in a web page and store that info in a way that it was searchable later on using a browser. I would be able to tell who changed what and when. I wanted to use SQL Server as it offers this feature, but only in SQL Server 2014. And since we recently upgraded to SQL 2012 it was cost prohibitive to upgrade to simply get the feature I wanted.

The first problem was to capture the information. So I began by writing a function that was called every time a page was posted back to the server. I would take the HTTP Request's Form object and save it to a log file. Since some pages can store quite a bit of info the log files grew rapidly, and often exceeded 250 Mb in size. Thus the second problem - trying to open a file this size using a Microsoft product is often futile and prone to crashes. 

The solution was to write a file parser that could split the single file into multiple files. It made sense to set this limit to 5Mb, since a browser can handle that file fairly well. And the language I chose was Perl. It is fast and lean. I created a service that runs on a given interval to review and parse files ready for splitting.